+++ test/testHelpers.ts
@@ -236,6 +236,7 @@ export class ServiceHandlersTestHarness {
   private readonly handlers = new Map<string, ServiceHandlerTestData>();
   private readonly allowedValues = new Map<string, string[]>();
   private readonly experimentalFeatures = new Set<string>();
+  private readonly serviceConfig = new Map<string, unknown>();
   readonly accessoryMock: MockProxy<BasicAccessory> & BasicAccessory;
 
   constructor() {
@@ -251,6 +252,11 @@ export class ServiceHandlersTestHarness {
     this.accessoryMock.isExperimentalFeatureEnabled
       .mockImplementation((feature: string): boolean => {
         return this.experimentalFeatures.has(feature.trim().toLocaleUpperCase());
+      });
+
+    this.accessoryMock.getServiceConfiguration
+      .mockImplementation((tag: string): unknown | undefined => {
+        return this.serviceConfig.get(tag);
       });
 
     this.accessoryMock.getOrAddService
@@ -294,6 +300,14 @@ export class ServiceHandlersTestHarness {
 
   clearExperimentalFeatureFlags(): void {
     this.experimentalFeatures.clear();
+  }
+
+  clearServiceConfigurations(): void {
+    this.serviceConfig.clear();
+  }
+
+  addServiceConfiguration(tag: string, value: unknown): void {
+    this.serviceConfig.set(tag, value);
   }
 
   private extractServiceId(id: ServiceIdentifier): string {